convert CheckoutPage to tsx
